Modularize and divide main.lua into multiple files
This commit moves some parts of `main.lua` into various other files
located in the `src/` directory. It also moves the libraries that were
located in root to `libs/`. This creater a nice distinction between the
different source files.

1. `main.lua` and `conf.lua`
   This is the _core_ of neko8.
2. `src/`
   Here are the various smaller modules that don't fit into _core_.
3. `libs/`
   Here are all the libraries from external sources, i.e. not made by us
   directly.
