--- api.lua
@@ -1,19 +1,0 @@
-function init() end
-function update() end
-function draw() end
-function text(text) end
-function keypressed(k, r) end
-function keyreleased(k) end
-
--- draw calls
-
-function pset(x, y) end
-function pget(x, y) end
-function spr(n, x, y, ...) end
-function sspr(n, ...) end
-
--- files
-
-function fileExists(name) end
-function getFileContents(name) end
-function writeFile(name, data) end
